削除処理にも制限

+++ app/Http/Controllers/Admin/BookController.php
@@ -84,6 +84,7 @@ class BookController extends Controller
     }
     public function destory(Book $book): RedirectResponse
     {
+        $this->authorize('delete', $book);
         // カスケードしているので
         $book->delete();
